src/frontend/requirements.txt
streamlit
streamlit_drawable_canvas
numpy
pillow
requests
